Move toolbar rendering into a sub-module.
